feat: GRC-20 v2 properties

+++ apps/web/core/utils/property/properties.ts
@@ -83,14 +83,14 @@ export function mapPropertyType(type: SwitchableRenderableType): PropertyTypeMap
         renderableTypeId: VIDEO_RENDERABLE_TYPE,
       };
     // GRC-20 v2 numeric types
-    case 'INT64':
+    case 'INTEGER':
       return {
-        baseDataType: 'INT64',
+        baseDataType: 'INTEGER',
         renderableTypeId: null,
       };
-    case 'FLOAT64':
+    case 'FLOAT':
       return {
-        baseDataType: 'FLOAT64',
+        baseDataType: 'FLOAT',
         renderableTypeId: null,
       };
     case 'DECIMAL':
@@ -157,8 +157,8 @@ export const typeToBaseDataType: Record<SwitchableRenderableType, DataType> = {
   IMAGE: 'RELATION',
   VIDEO: 'RELATION',
   // GRC-20 v2 numeric types
-  INT64: 'INT64',
-  FLOAT64: 'FLOAT64',
+  INTEGER: 'INTEGER',
+  FLOAT: 'FLOAT',
   DECIMAL: 'DECIMAL',
   // GRC-20 v2 boolean type
   BOOL: 'BOOL',
